backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
sympy==1.12
python-dotenv==1.0.1
websockets==12.0
